Pull any classes that were nested inside a file with other code out into their own files
